refactor: API별 타입 분리 #1
- ProductDelete, ProductDetailAsk, ProductListAsk
