--- OpenMarket/OpenMarket/NetWorking/NetWorkingType/ProductPageHTTPResponse.swift
@@ -1,23 +1,0 @@
-import Foundation
-// 2-5 상품 list조회 , 
-struct ProductPageHTTPResponse: Codable {
-    let pageNo: Int
-    let itemsPerPage: Int
-    let totalCount: Int
-    let offset: Int
-    let limit: Int
-    let lastPage: Int
-    let hasNext: Bool
-    let hasPrev: Bool
-    let pages: [ProductHTTPResponse]
-    
-    enum CodingKeys: String, CodingKey {
-        case offset, limit, pages
-        case pageNo = "page_no"
-        case itemsPerPage = "items_per_page"
-        case totalCount = "total_count"
-        case lastPage = "last_page"
-        case hasNext = "has_next"
-        case hasPrev = "has_prev"
-    }
-}
